fixed #43
[ # ] get_from_code & get_from_id now pass linkInfoV2 as root JSON key into the deserializer
[ # ] LastMessageSummary.MediaType is now an int? instead of string
[ # ] WalletInfo.TotalCoins is now a float? instead of int? because Amino API is weird (there literally is a totalCoinsFloat value????)
[ # ] SubClient.send_activity_time now sends optInAdsFlags instead of OptInAdsFlags

+++ Amino.NET/Objects/WalletInfo.cs
@@ -10,7 +10,7 @@ namespace Amino.Objects
         [JsonPropertyName("adsEnabled")]public bool? AdsEnabled { get; set; }
         [JsonPropertyName("adsVideoStats")]public string AdsVideoStats { get; set; }
         [JsonPropertyName("adsFlag")]public string AdsFlag { get; set; }
-        [JsonPropertyName("totalCoins")]public int? TotalCoins { get; set; }
+        [JsonPropertyName("totalCoins")]public float? TotalCoins { get; set; }
         [JsonPropertyName("businessCoinsEnabled")]public bool? BusinessCoinsEnabled { get; set; }
         [JsonPropertyName("totalBusinessCoins")]public int? TotalBusinessCoins { get; set; }
         [JsonPropertyName("totalBusinessCoinsFloat")]public float? TotalBusinessCoinsFloat { get; set; }
